+++ R/mirai-package.R
@@ -49,7 +49,7 @@
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
 #' @importFrom nanonext call_aio context is_error_value msleep random recv
-#'     request send send_aio socket stop_aio unresolved
+#'     recv_aio request send socket stop_aio unresolved
 #'
 #' @docType package
 #' @name mirai-package
@@ -68,4 +68,8 @@ NULL
                   Linux = "abstract://n%.f",
                   Windows = "ipc://n%.f",
                   "ipc:///tmp/n%.f")
+
+.__scm__. <- as.raw(c(0x58, 0x0a, 0x00, 0x00, 0x00, 0x03, 0x00, 0x04, 0x02,
+                      0x01, 0x00, 0x03, 0x05, 0x00, 0x00, 0x00, 0x00, 0x05,
+                      0x55, 0x54, 0x46, 0x2d, 0x38, 0x00, 0x00, 0x00, 0xfc))